requirements.txt
Django==4.2.16
django-redis==5.4.0
psycopg==3.2.3
djangorestframework==3.15.2
drf-spectacular==0.27.2
djangorestframework-simplejwt==5.3.1
